Lia_Leroy-Merlin/requirements.txt
streamlit==1.29.0
transformers==4.36.2
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
pillow==10.2.0
langchain==0.0.350
langchain-community==0.0.10
chromadb==0.4.22
python-docx==1.0.1
PyPDF2==3.0.1 